Update mmosga.sh
Make messages better
Make script compatible with Intel Macs and Hackintoshs (with dynamic
homebrew prefix)
Add eza
Make Ghostty theme selection better

+++ mmosga.sh
@@ -17,6 +17,8 @@ if ! gum confirm "Do you want to proceed?" ; then
   exit 1
 fi
 
+brewprefix="$(brew --prefix)"
+
 clear
 # Dock
 defaults write com.apple.dock magnification -bool true # enable magnification
@@ -24,7 +26,7 @@ defaults write com.apple.dock largesize -int 80 # set magnification level
 defaults write com.apple.dock autohide -bool true # enable autohide
 defaults write com.apple.dock autohide-time-modifier -float 0.5 # speed up autohide animation
 defaults write com.apple.dock autohide-delay -float 0 # remove autohide delay
-echo "Dock settings are finished."
+echo "Dock settings are finished..."
 
 # Finder
 defaults write com.apple.finder AppleShowAllFiles -bool true # enable hidden files
@@ -35,28 +37,28 @@ defaults write com.apple.finder NewWindowTarget -string "PfHm" # set new finder 
 defaults write com.apple.finder NewWindowTargetPath -string "file://${HOME}/" # set Finder home path to user's home folder
 defaults write com.apple.finder QLEnableTextSelection -bool true # enable text selection in quick look
 defaults write com.apple.desktopservices DSDontWriteNetworkStores -bool true # don't store DS_Store files on network
-echo "Finder settings are finished."
+echo "Finder settings are finished..."
 
 # Keyboard
 defaults write NSGlobalDomain NSAutomaticSpellingCorrectionEnabled -bool false # disable automatic spelling
 defaults write NSGlobalDomain NSAutomaticCapitalizationEnabled -bool false # disable automatic capitalization
 defaults write NSGlobalDomain NSAutomaticPeriodSubstitutionEnabled -bool false # disable automatic period substitution
 defaults write NSGlobalDomain NSAutomaticQuoteSubstitutionEnabled -bool false # disable automatic quote substitution defaults write NSGlobalDomain NSAutomaticDashSubstitutionEnabled -bool false # disable automatic dash substitution
-echo "Keyboard settings are finished."
+echo "Keyboard settings are finished..."
 
 # Trackpad
 defaults write NSGlobalDomain com.apple.swipescrolldirection -bool false # disable natural scrolling in trackpad
-echo "Trackpad settings are finished."
+echo "Trackpad settings are finished..."
 
 # Accent color (purple)
 defaults write -g AppleAccentColor -int 5
 defaults write -g AppleColorPreferences -dict AccentColor -int 5
 defaults write -g AppleHighlightColor -string "0.968627 0.831373 1.000000 Purple"
-echo "Accent color has been set up to purple."
+echo "Accent color has been set up to purple..."
 
 # Quarantine settings
 defaults write com.apple.LaunchServices LSQuarantine -bool false # less Are you sure? prompts
-echo "Quarantine settings are finished."
+echo "Quarantine settings are finished..."
 
 # Kill services
 killall Dock SystemUIServer Finder TextEdit
@@ -83,7 +85,7 @@ fi
 if ! pkgutil --pkg-info com.apple.pkg.RosettaUpdateAuto >/dev/null && gum confirm "Do you want to install Rosetta 2?" ; then
   softwareupdate --install-rosetta --agree-to-license
 else
-  echo "Skipping Rosetta 2 because you already have it."
+  echo "Skipping Rosetta 2..."
 fi
 
 # Neovim
@@ -112,7 +114,7 @@ if command -v nvim >/dev/null; then
         ;;
     esac
   else
-    echo "Skipping Neovim config section because you already have a Neovim config."
+    echo "Skipping Neovim config section..."
   fi
 fi
 
@@ -131,7 +133,7 @@ if command -v emacs >/dev/null; then
         ;;
     esac
   else
-    echo "Skipping Emacs config section because you already have a Emacs config."
+    echo "Skipping Emacs config section..."
   fi
 fi
 
@@ -139,12 +141,12 @@ fi
 if command -v zsh >/dev/null; then
   if [ ! -f "$HOME/.zshrc" ] && gum confirm "Do you want to configure Zsh?"; then
     if gum confirm "Do you want to use zsh-syntax-highlighting?"; then
-      brew install zsh-syntax-highlighting
-      echo "source /opt/homebrew/share/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> ~/.zshrc # enable autosyntax
+      [[ ! -f "$brewprefix/share/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" ]] && brew install zsh-syntax-highlighting
+      echo "source $brewprefix/share/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> ~/.zshrc # enable autosyntax
     fi
     if gum confirm "Do you want to use zsh-autosuggestions?"; then
-      brew install zsh-autosuggestions
-      echo "source /opt/homebrew/share/zsh-autosuggestions/zsh-autosuggestions.zsh" >> ~/.zshrc # enable autosuggestions
+      [[ ! -f "$brewprefix/share/zsh-autosuggestions/zsh-autosuggestions.zsh" ]] && brew install zsh-autosuggestions
+      echo "source $brewprefix/share/zsh-autosuggestions/zsh-autosuggestions.zsh" >> ~/.zshrc # enable autosuggestions
     fi
     if gum confirm "Do you want to use a custom prompt?"; then
       echo "export PS1='%B%F{Cyan}%~%b%F{white} $ '" >> ~/.zshrc # a better prompt
@@ -154,8 +156,14 @@ if command -v zsh >/dev/null; then
       echo "compinit" >> ~/.zshrc # modern zsh autocomplete system
       echo "zstyle ':completion:*' menu select" >> ~/.zshrc # use arrow keys
     fi
+    if gum confirm "Do you want to use eza (ls alternative)?"; then
+      command -v eza >/dev/null || brew install eza || true
+      echo 'alias ls="eza"' >> ~/.zshrc
+      echo 'alias la="eza -a"' >> ~/.zshrc
+      echo 'alias tree="eza -T"' >> ~/.zshrc
+    fi
   else
-    echo "Skipping Zsh config section because you already have a Zsh config."
+    echo "Skipping Zsh config section..."
   fi
 fi
 
@@ -165,10 +173,10 @@ if command -v ghostty >/dev/null; then
     [[ ! -d ~/.config/ghostty ]] && mkdir -p ~/.config/ghostty # create config folder if doesn't exist
     echo "background-opacity = $(gum input --placeholder 'Transparency (between 0.0 and 1.0)')" >> ~/.config/ghostty/config.ghostty
     gum confirm "Do you want to use Option key as Alt?" && echo "macos-option-as-alt = left" >> ~/.config/ghostty/config.ghostty
-    scheme=$(ghostty +list-themes | sed 's/ (.*)//' | gum choose --header "Select a color scheme")
+    scheme=$(ghostty +list-themes | sed -E 's/ \(resources\)$//' | gum choose --header "Select a color scheme")
     echo "theme = $scheme" >> ~/.config/ghostty/config.ghostty
   else
-    echo "Skipping Ghostty config section because you already have a Ghostty config."
+    echo "Skipping Ghostty config section..."
   fi
 fi
